Done experiment

+++ src/model.cxx
@@ -18,8 +18,8 @@ model model::model_from_file(string s) {
     return ret;
 }
 
-const map<symbol, pair<float, context>> model::cand_and_p(context c) const {
-    map<symbol, pair<float, context>> ret;
+const map<symbol, pair<double, context>> model::cand_and_p(context c) const {
+    map<symbol, pair<double, context>> ret;
     for (int i = 0; i < cnt; ++i)
         if (buf[c][i])
             ret[i] = make_pair(buf[c][i], i);
